Fixes and code improvement.

+++ src/main/java/PennyPincher/service/expenses/ExpenseServiceImpl.java
@@ -3,30 +3,26 @@ package PennyPincher.service.expenses;
 import PennyPincher.entity.Expense;
 import PennyPincher.entity.Payoff;
 import PennyPincher.entity.User;
-import PennyPincher.repository.EventRepository;
+import PennyPincher.exception.ExpenseNotFoundException;
 import PennyPincher.repository.ExpenseRepository;
-import PennyPincher.repository.UserRepository;
 import lombok.AllArgsConstructor;
-import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @AllArgsConstructor
 @Service
 public class ExpenseServiceImpl implements ExpenseService {
+
     private ExpenseRepository expenseRepository;
 
     @Override
     public Expense findById(Integer expenseId) {
         return expenseRepository.findById(expenseId)
-                .orElseThrow(() -> new ResourceNotFoundException("Expense not found with id: " + expenseId));
+                .orElseThrow(() -> new ExpenseNotFoundException("Expense not found with ID: " + expenseId));
     }
 
     @Override
@@ -35,15 +31,27 @@ public class ExpenseServiceImpl implements ExpenseService {
     }
 
     @Override
+    public void deleteByEventId(Integer eventId) {
+        expenseRepository.deleteById(eventId);
+    }
+
+    @Override
     public void deleteById(Integer expenseId) {
+        if (!expenseRepository.existsById(expenseId)) {
+            throw new ExpenseNotFoundException("Expense not found with ID: " + expenseId);
+        }
         expenseRepository.deleteById(expenseId);
     }
 
     @Override
     public List<Expense> findExpensesForGivenEvent(Integer eventId) {
-        return expenseRepository.findAll().stream()
-                .filter(expense -> eventId.equals(expense.getEvent().getId()))
-                .collect(Collectors.toList());
+        Optional<List<Expense>> expensesOptional = expenseRepository.findByEventId(eventId);
+        return expensesOptional.orElse(Collections.emptyList());
+    }
+
+    @Override
+    public Optional<Expense> findByExpenseNameAndEventId(String expenseName, Integer eventId) {
+        return expenseRepository.findByNameAndEventId(expenseName, eventId);
     }
 
     @Override
@@ -55,39 +63,29 @@ public class ExpenseServiceImpl implements ExpenseService {
 
     @Override
     public Map<Integer, BigDecimal> mapUserToCost(Expense expense) {
-        Map<Integer, BigDecimal> mapUserPerCost = new HashMap<>();
-        expense.getParticipants().forEach(participant -> mapUserPerCost.put(participant.getId(),
-                sumParticipantCosts(expense, participant)
-        ));
-        return mapUserPerCost;
+        return expense.getParticipants().stream()
+                .collect(Collectors.toMap(
+                        User::getId,
+                        participant -> sumParticipantCosts(expense, participant)
+                ));
     }
 
     @Override
     public Map<Integer, BigDecimal> mapUserToPayoffAmount(Expense expense) {
-        Map<Integer, BigDecimal> mapUserPerPayoffsAmount = new HashMap<>();
-        expense.getParticipants().forEach(participant -> mapUserPerPayoffsAmount.put(participant.getId(),
-                sumParticipantPayoffs(expense, participant)
-        ));
-        return mapUserPerPayoffsAmount;
+        return expense.getParticipants().stream()
+                .collect(Collectors.toMap(
+                        User::getId,
+                        participant -> sumParticipantPayoffs(expense, participant)
+                ));
     }
 
     @Override
     public Map<Integer, BigDecimal> mapUserToBalance(Expense expense) {
-        Map<Integer, BigDecimal> mapUserPerBalance = new HashMap<>();
-        expense.getParticipants().forEach(participant -> mapUserPerBalance.put(participant.getId(),
-                calculateParticipantBalance(expense, participant)
-        ));
-        return mapUserPerBalance;
-    }
-
-    @Override
-    public void deleteByEventId(Integer eventId) {
-        expenseRepository.deleteById(eventId);
-    }
-
-    @Override
-    public Expense findByExpenseNameAndEventId(String expenseName, Integer eventId) {
-        return expenseRepository.findByNameAndEventId(expenseName, eventId);
+        return expense.getParticipants().stream()
+                .collect(Collectors.toMap(
+                        User::getId,
+                        participant -> calculateParticipantBalance(expense, participant)
+                ));
     }
 
     private BigDecimal sumParticipantPayoffs(Expense expense, User participant) {
